jsq changed to real harvest kvp interfaces

+++ openwhisk-harv-vm-cgroup-azure-distributed-jsq/common/scala/src/main/scala/org/apache/openwhisk/core/connector/Message.scala
@@ -171,7 +171,7 @@ object AcknowledegmentMessage extends DefaultJsonProtocol {
 }
 
 // yanqi, add rsc info to PingMessage, core/cpu number and memory (in MB)
-case class PingMessage(instance: InvokerInstanceId, cpu: Int, memory: Int) extends Message {
+case class PingMessage(instance: InvokerInstanceId, cpu: Double, memory: Long) extends Message {
   override def serialize = PingMessage.serdes.write(this).compactPrint
 }
